:pushpin: Trim versions to just two digits
This will allow users to do something like the following:

```python
self.tool_requires(
  f"arm-gnu-toolchain/{self.settings.compiler.version}")
```

Note that this has some issues in that, it becomes `None` if the field
is not set. What would be better is if it was checked to be valid like
so:

```python
@property
def _compiler_version(self):
    SUPPORTED_COMPILER_VERSIONS = ["11.3", "12.2", "12.3"]

    if self.settings.compiler.version not in SUPPORTED_COMPILER_VERSIONS:
        raise ConanInvalidConfiguration(
            "The GCC compiler version must be one of these " +
            f"{SUPPORTED_COMPILER_VERSIONS}, provided version: " +
            f"'{self.settings.compiler.version}'")

    return str(self.settings.compiler.version)
```

Demos now build demo application to prove that the package works as
test_package is not an adequate means to test this.